fix(float): f32 constant folding + f32 vreg tracking in x86_64
Two fixes for f32 float codegen:

1. opt.rs try_fold_binop: for F32 operations, narrow f64 bits to f32
   FIRST (f64::from_bits(lhs) as f32 as f64), then do arithmetic.
   Previously the code did f32::from_bits(lhs as u32) which reinterpreted
   the LOW 32 bits of f64 bits as f32 — wrong because f64 bits for 3.0
   (0x4008000000000000) have 0x00000000 in the low 32 bits (= 0.0 as f32).
   This caused 3.0+2.0 to fold to 0.0+0.0 = 0.0, and 0.0/0.0 = NaN.

2. x86_64 stack_slot_isel.rs: infer_fp_vregs now returns a SECOND set
   (fp_vregs_f32) tracking vregs that specifically hold f32 (not f64).
   The is_f64 check now uses: matches!(ty, Some(F64)) || (ty.is_none()
   && !fp_vregs_f32.contains(&dst_id)). Previously is_f64 defaulted to
   true when ty was None, causing f32 operations to use f64 instructions.

3. x86_64 BinOp Add path: added load_value_f32 helper that narrows f64
   immediate bits to f32 bits before loading into XMM (only for the Add
   path so far; Sub/Mul/Div/Cmp paths need the same fix).

Result: f32_to_f64_widen now PASSES. f32_arith and f64_to_f32_narrow
still fail (need load_value_f32 applied to all FP paths + ty propagation
fix in the IR builder).

+++ src/codegen/src/opt.rs
@@ -263,8 +263,21 @@ fn try_fold_binop(op: BinOpKind, lhs: i64, rhs: i64, ty: Option<&IRType>) -> Opt
     // G7: if the operand type is F32/F64, fold using f64 arithmetic.
     if let Some(IRType::F32) | Some(IRType::F64) = ty {
         let is_f64 = matches!(ty, Some(IRType::F64));
-        let lf = if is_f64 { f64::from_bits(lhs as u64) } else { f32::from_bits(lhs as u32) as f64 };
-        let rf = if is_f64 { f64::from_bits(rhs as u64) } else { f32::from_bits(rhs as u32) as f64 };
+        // All float literals in VUMA are f64 at the AST level, stored as
+        // f64 bits in i64 immediates. For f32 operations, we must narrow
+        // the f64 to f32 first (NOT reinterpret the low 32 bits as f32,
+        // which would give wrong values since f64 bits ≠ f32 bits).
+        let lf = if is_f64 {
+            f64::from_bits(lhs as u64)
+        } else {
+            // Narrow f64 to f32, then back to f64 for arithmetic.
+            f64::from_bits(lhs as u64) as f32 as f64
+        };
+        let rf = if is_f64 {
+            f64::from_bits(rhs as u64)
+        } else {
+            f64::from_bits(rhs as u64) as f32 as f64
+        };
         let result: f64 = match op {
             BinOpKind::Add => lf + rf,
             BinOpKind::Sub => lf - rf,
